--- a/warc-indexer/src/main/java/uk/bl/wa/analyser/payload/PDFAnalyser.java
+++ b/warc-indexer/src/main/java/uk/bl/wa/analyser/payload/PDFAnalyser.java
@@ -38,6 +38,7 @@ import uk.bl.wa.solr.SolrFields;
 import uk.bl.wa.solr.SolrRecord;
 
 import com.typesafe.config.Config;
+import uk.bl.wa.util.Instrument;
 
 /**
  * @author anj
@@ -58,6 +59,7 @@ public class PDFAnalyser extends AbstractPayloadAnalyser {
 	@Override
 	public void analyse(ArchiveRecordHeader header, InputStream tikainput,
 			SolrRecord solr) {
+        final long start = System.nanoTime();
 		Metadata metadata = new Metadata();
 		metadata.set(Metadata.RESOURCE_NAME_KEY, header.getUrl());
 		ParseRunner parser = new ParseRunner(app, tikainput, metadata, solr);
@@ -96,6 +98,7 @@ public class PDFAnalyser extends AbstractPayloadAnalyser {
 				solr.addField(SolrFields.PDFA_ERRORS, error);
 			}
 		}
+        Instrument.timeRel("PDFAnalyzer.analyze", start);
 	}
 
 }