--- a/warc-indexer/src/main/java/uk/bl/wa/analyser/payload/WARCPayloadAnalysers.java
+++ b/warc-indexer/src/main/java/uk/bl/wa/analyser/payload/WARCPayloadAnalysers.java
@@ -40,6 +40,7 @@ import uk.bl.wa.nanite.droid.DroidDetector;
 import uk.bl.wa.solr.SolrFields;
 import uk.bl.wa.solr.SolrRecord;
 import uk.bl.wa.solr.TikaExtractor;
+import uk.bl.wa.util.Instrument;
 import uk.gov.nationalarchives.droid.command.action.CommandExecutionException;
 
 import com.google.common.base.Splitter;
@@ -109,7 +110,8 @@ public class WARCPayloadAnalysers {
 	
 	public void analyse(ArchiveRecordHeader header, InputStream tikainput, SolrRecord solr) {
 		log.debug("Analysing "+header.getUrl());
-		
+
+        final long tikaStart = System.nanoTime();
 		// Analyse with tika:
 		try {
 			if( passUriToFormatTools ) {
@@ -120,7 +122,9 @@ public class WARCPayloadAnalysers {
 		} catch( Exception i ) {
 			log.error( i + ": " + i.getMessage() + ";tika; " + header.getUrl() + "@" + header.getOffset() );
 		}
+        Instrument.timeRel("WARCPayloadAnalyzers.analyze#tika", tikaStart);
 
+        final long firstBytesStart = System.nanoTime();
 		// Pull out the first few bytes, to hunt for new format by magic:
 		try {
 			tikainput.reset();
@@ -141,9 +145,11 @@ public class WARCPayloadAnalysers {
 		} catch( Exception i ) {
 			log.error( i + ": " + i.getMessage() + ";ffb; " + header.getUrl() + "@" + header.getOffset() );
 		}
+        Instrument.timeRel("WARCPayloadAnalyzers.analyze#firstbytes", firstBytesStart);
 
 		// Also run DROID (restricted range):
 		if( dd != null && runDroid == true ) {
+            final long droidStart = System.nanoTime();
 			try {
 				tikainput.reset();
 				// Pass the URL in so DROID can fall back on that:
@@ -161,6 +167,7 @@ public class WARCPayloadAnalysers {
 				// Note that DROID complains about some URLs with an IllegalArgumentException.
 				log.error( i + ": " + i.getMessage() + ";dd; " + header.getUrl() + " @" + header.getOffset() );
 			}
+            Instrument.timeRel("WARCPayloadAnalyzers.analyze#droid", droidStart);
 		}
 		
 		try {