--- a/warc-indexer/src/main/java/uk/bl/wa/extract/LanguageDetector.java
+++ b/warc-indexer/src/main/java/uk/bl/wa/extract/LanguageDetector.java
@@ -116,7 +116,7 @@ public class LanguageDetector {
                 return this.getLangdetectLanguage(text);
             return null;
         } finally {
-            Instrument.timeRel("LanguageAnalyzer#total", "LanguageDetector.detectLanguage#ld", start);
+            Instrument.timeRel("LanguageAnalyzer#total", "LanguageDetector.detectLanguage#ld", startLD);
         }
 	}
 	/**