… warc-indexer/src/main/java/uk/bl/wa/solr/SolrRecord.java
… warc-indexer/src/main/java/uk/bl/wa/util/Instrument.java
… warc-indexer/src/main/resources/reference.conf
… warc-indexer/src/test/java/uk/bl/wa/solr/TikaExtractorTest.java
… warc-indexer/src/test/java/uk/bl/wa/util/InstrumentTest.java