This pull is primarily about performance instrumentation: The class Instrument was added for lightweight tracking of where the time is spend (sample output below). The final report is currently only triggered by the `WARCIndexerCommand` with `Instrument.log(true)`.

Eye opener: Detecting language on our (admittedly tiny) sample took 1/5th of the total time.

```
WARC Indexer Finished in 180.572 seconds.
2015-02-12 15:43:12 INFO  Instrument:166 - Performance statistics
SolrRecord.removeControlCharacters#total(#=253778, time=3003.60ms, avg=84.49#/ms 0.01ms/#)
  SolrRecord.sanitiseUTF8(#=253778, time=961.46ms, avg=263.95#/ms 0.00ms/#)
WARCIndexerCommand.main#total(#=1, time=180577.93ms, avg=0.00#/ms 180577.93ms/#)
  WARCIndexerCommand.parseWarcFiles#startup(#=1, time=4681.03ms, avg=0.00#/ms 4681.03ms/#)
  WARCIndexerCommand.parseWarcFiles#solrdocCreation(#=2973, time=172194.94ms, avg=0.02#/ms 57.92ms/#)
    WARCIndexer.extract#total(#=2973, time=171782.62ms, avg=0.02#/ms 57.78ms/#)
      TextAnalyzers#total(#=2973, time=51045.44ms, avg=0.06#/ms 17.17ms/#)
        LanguageAnalyzer#total(#=2762, time=49251.77ms, avg=0.06#/ms 17.83ms/#)
          LanguageDetector.detectLanguage#li(#=2762, time=37463.25ms, avg=0.07#/ms 13.56ms/#)
          LanguageDetector.detectLanguage#ld(#=2762, time=11740.58ms, avg=0.24#/ms 4.25ms/#)
        FuzzyHashAnalyzer(#=2762, time=1411.92ms, avg=1.96#/ms 0.51ms/#)
        PostcodeAnalyzer(#=2762, time=355.36ms, avg=7.77#/ms 0.13ms/#)
      WARCIndexer.extract#hashstreamwrap(#=2973, time=6397.88ms, avg=0.46#/ms 2.15ms/#)
      WARCIndexer.extract#archeaders(#=3377, time=971.33ms, avg=3.48#/ms 0.29ms/#)
      WARCIndexer.extract#analyzetikainput(#=2973, time=113377.54ms, avg=0.03#/ms 38.14ms/#)
        WARCPayloadAnalyzers.analyze#total(#=2973, time=113366.16ms, avg=0.03#/ms 38.13ms/#)
          HTMLAnalyzer.analyze#total(#=2730, time=34759.37ms, avg=0.08#/ms 12.73ms/#)
            HTMLAnalyzer.analyze#parser(#=2730, time=30107.75ms, avg=0.09#/ms 11.03ms/#)
          WARCPayloadAnalyzers.analyze#droid(#=2973, time=16764.40ms, avg=0.18#/ms 5.64ms/#)
          XMLAnalyzer.analyze(#=32, time=34.01ms, avg=0.94#/ms 1.06ms/#)
          WARCPayloadAnalyzers.analyze#tikasolrextract(#=2973, time=54007.04ms, avg=0.06#/ms 18.17ms/#)
          ImageAnalyzer.analyze#facesanddominant(#=13, time=7660.43ms, avg=0.00#/ms 589.26ms/#)
          WARCPayloadAnalyzers.analyze#firstbytes(#=2973, time=125.02ms, avg=23.78#/ms 0.04ms/#)
```