@anjackson 2015-02-16
The warc-indexer/src/main/java/uk/bl/wa/util/Instrument.java and the InstrumentTest.java files have two headers with two different licenses. Are you okay with these falling under the license of the overall project?