@tokee 2015-02-16
Whoops. Yes, no problem. That's what I get for using class templates with license in my IDE.